+++ DemoApp/AppDelegate.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
 import Atlantis
@@ -12,9 +12,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         _ application: UIApplication,
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
-        // Override point for customization after application launch.
-        Atlantis.start()
-        return true
+        // Uncomment this to be able to proxy WebSocket events from Proxyman.app
+        // Atlantis.start()
+        true
     }
 
     func application(